Split BaseContext and BindingContext in SQLStatementContext
